Update total overunder to map to float type.

+++ src/config/mappings.py
@@ -107,7 +107,7 @@ betting_dtype_mapping = {
     "game_id": 'object',
     "last_update": 'float',  # Update to datetime after mapping
     "sportsbook_id": 'int64', 
-    "total_over_under": 'int64',
+    "total_over_under": 'float',
     "over_odds": 'object',
     "under_odds": 'object',
     "home_spread": 'object',
